+++ src/main/kotlin/hu/raven/puppet/logic/step/orderpopulationbycost/OrderPopulationByCost.kt
@@ -1,5 +1,6 @@
 package hu.raven.puppet.logic.step.orderpopulationbycost
 
+import hu.akos.hollo.szabo.math.vector.FloatVector.Companion.length
 import hu.raven.puppet.logic.operator.calculatecost.CalculateCost
 import hu.raven.puppet.logic.step.EvolutionaryAlgorithmStep
 import hu.raven.puppet.model.state.EvolutionaryAlgorithmState
@@ -16,6 +17,6 @@ class OrderPopulationByCost(
                 specimen.cost = calculateCostOf(specimen)
             }
 
-        sortActiveBy { it.costOrException().length() }
+        sortActiveBy { it.costOrException().length().toFloat() }
     }
 }
